09-Audio-Chat/requirements.txt
azure-identity
azure-ai-projects
openai>=1.0.0
python-dotenv
numpy
wave